% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elicitBivariate.R
\name{elicitBivariate}
\alias{elicitBivariate}
\alias{elicitConcProb}
\title{Elicit a bivariate distribution using a Gaussian copula}
\usage{
elicitBivariate()
}
\value{
A list, with two objects of class \code{elicitation}, and the 
elicited concordance probability. See \code{\link{fitdist}} for details.
}
\description{
Opens up a web browser (using the shiny package), from which you can specify
judgements, fit distributions, plot the fitted density functions, and plot samples 
from the joint distributions. A joint distribution is constructed using a Gaussian
copula, whereby the correlation parameter is determined via the elicitation of a 
concordance probability (a probability that the two uncertain quantities are either
both greater than their medians, or both less than their medians.)
}
\details{
Click on the "Help" tab for instructions. Click the "Quit" button to exit the app and return
the results from the \code{fitdist} command. Click "Download report" to generate a report
of all the fitted distributions for each uncertain quantity, and "Download sample" to
generate a csv file with a sample from the joint distribution.
}
\examples{

\dontrun{

elicitBivariate()

}
}
\author{
Jeremy Oakley <j.oakley@sheffield.ac.uk>
}
