% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elicitDirichlet.R
\name{elicitDirichlet}
\alias{elicitDirichlet}
\title{Elicit a Dirichlet distribution interactively}
\usage{
elicitDirichlet()
}
\value{
The parameters of the fitted Dirichlet distribution, which are 
returned once the Quit button has been clicked.
}
\description{
Opens up a web browser (using the shiny package), from which you can elicit a
Dirichlet distribution
}
\details{
Click on the "Help" tab for instructions. Click the "Quit" button to exit the app and return
the results from the \code{fitdist} command. Click "Download report" to generate a report
of all the fitted distributions.
}
\examples{

\dontrun{

elicit()

}
}
\author{
Jeremy Oakley <j.oakley@sheffield.ac.uk>
}
