% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elicitExtension.R
\name{elicitExtension}
\alias{elicitExtension}
\title{Elicitation with the extension method}
\usage{
elicitExtension()
}
\value{
A list, with two objects of class \code{elicitation}.
 See \code{\link{fitdist}} for details.
}
\description{
Opens up a web browser (using the shiny package), from which you can specify
judgements, fit distributions, and produce various plots. Judgements are 
specified for the distribution of the conditioning variable Y, the median 
function (median of X given Y), and the distribution of X given that Y takes
its median value. Plots are provided for the two elicited distributions, the
median function, the conditional distribution of X for any specified Y, and
the marginal distribution of X.
}
\details{
Click the "Quit" button to exit the app and return
the results from the \code{fitdist} command. Click "Download report" to generate a report
of all the fitted distributions for each uncertain quantity, and "Download sample" to
generate a csv file with a sample from the marginal distribution of X.
}
\examples{

\dontrun{

elicitExtension()

}
}
\author{
Jeremy Oakley <j.oakley@sheffield.ac.uk>
}
