% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elicitMixture.R
\name{elicitMixture}
\alias{elicitMixture}
\title{Elicit a mixture distribution using the extension method}
\usage{
elicitMixture()
}
\value{
When the Quit button is clicked, a list, with elements
\item{fit}{an object of class \code{elicitation}. See \code{\link{fitdist}} for details.}
\item{extensionProbs}{the probability mass function for the extension variable.}
}
\description{
Opens up a web browser (using the shiny package), from which you can specify
judgements, fit distributions and plot the fitted density function.
}
\details{
Click the "Quit" button to exit the app and return
the fitted distributions. Click "Download report" to generate a report
of all the fitted distributions.
}
\examples{

\dontrun{

elicitMixture()

}
}
\author{
Jeremy Oakley <j.oakley@sheffield.ac.uk>
}
