% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elicitQuartiles.R
\name{elicitQuartiles}
\alias{elicitQuartiles}
\title{Elicit judgements and fit distributions interactively using the quartile method}
\usage{
elicitQuartiles()
}
\description{
Opens up a web browser (using the shiny package), from which you can specify
the quartiles, fit distributions and plot the fitted density functions with
additional feedback.
}
\details{
Parameter limits determine which distributions can be fitted. Finite
lower limits are needed for the gamma, lognormal and log-t distributions,
and both limits must be finite for to fit a beta distribution. If a
histogram is fitted without specifying finite limits, endpoints are chosen
based on fitting a normal distribution.

Press Esc in the R console window to exit the elicitation session.
}
\examples{

\dontrun{

elicitQuartiles()

}
}
\author{
Jeremy Oakley <j.oakley@sheffield.ac.uk>
}
