% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elicitSHELF.R
\name{elicitSHELF}
\alias{elicitSHELF}
\title{SHELF Elicitation}
\usage{
elicitSHELF()
}
\description{
Opens up a web browser in which you can implement the general SHELF protocol of individual elicitation
followed by group discussion and elicitation of a single set of Rational Impartial Observer (RIO) probabilities
}
\examples{

\dontrun{


elicitSHELF()

}
}
\author{
Jeremy Oakley <j.oakley@sheffield.ac.uk>
}
