% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elicitTertiles.R
\name{elicitTertiles}
\alias{elicitTertiles}
\title{Elicit judgements and fit distributions interactively using the tertile method}
\usage{
elicitTertiles()
}
\value{
An object of class \code{elicitation}, which is returned once the 
Finish button has been clicked. See \code{\link{fitdist}} for details.
}
\description{
Opens up a web browser (using the shiny package), from which you can specify
the median and tertiles, fit distributions and plot the fitted density functions with
additional feedback.
}
\details{
Parameter limits determine which distributions can be fitted. Finite
lower limits are needed for the gamma, lognormal and log-t distributions,
and both limits must be finite for to fit a beta distribution. If a
histogram is fitted without specifying finite limits, endpoints are chosen
based on fitting a normal distribution.


Click the Finish button to quit the elicitation session.
}
\examples{

\dontrun{

elicitTertiles()

}
}
\author{
Jeremy Oakley <j.oakley@sheffield.ac.uk>
}
