% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateReport.R
\name{generateReport}
\alias{generateReport}
\title{Generate a report to show the fitted distributions}
\usage{
generateReport(
  fit,
  output_format = "html_document",
  sf = 3,
  expert = 1,
  view = TRUE,
  clean = TRUE
)
}
\arguments{
\item{fit}{An object of class \code{'elicitation'}.}

\item{output_format}{the output format for the document. One of \code{"html_document"},
\code{"pdf_document"} (requires LaTeX to be installed), or \code{"word_document"}
(requires Word to be installed).}

\item{sf}{number of significant figures to be displayed for the fitted parameters.}

\item{expert}{if the \code{fit} object contains judgements from multiple experts, the 
single expert's distributions to be displayed.}

\item{view}{set to \code{TRUE} to open the document after it has been compiled.}

\item{clean}{set to \code{TRUE} to clean intermediate files that are created during rendering.}
}
\description{
Renders an Rmarkdown document to display the density function of each fitted distribution, 
the parameter values, and the R command required to sample from each distribution.
}
\examples{
\dontrun{
# One expert, with elicited probabilities
# P(X<20)=0.25, P(X<30)=0.5, P(X<50)=0.75
# and X>0.
v <- c(20,30,50)
p <- c(0.25,0.5,0.75)
myfit <- fitdist(vals=v, probs=p, lower=0)

generateReport(myfit)
}
}
