% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSurvivalTable.R
\name{makeSurvivalTable}
\alias{makeSurvivalTable}
\title{Tabulate Summary Data for Survival Extrapolation}
\usage{
makeSurvivalTable(
  survDF,
  breakTime,
  truncationTime,
  timeUnit,
  useWeights = FALSE,
  dp = 2
)
}
\arguments{
\item{survDF}{}

\item{breakTime}{}

\item{truncationTime}{}

\item{timeUnit}{}

\item{dp}{}
}
\description{
Tabulates the Kaplan Meier survivor function and within interval hazard at discrete equally spaced time points t_1,...,t_n
"Within interval hazard" is defined as (1-S(t_{n+1})) / S_(t_n), using the Kaplan Meier estimate of S().
The table is intended to be included on a summary sheet provided to experts when eliciting judgements about
extrapolated survival probabilities.
}
