% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyLP.R
\name{plotWeightLP}
\alias{plotWeightLP}
\title{Plot a weighted linear pool}
\usage{
plotWeightLP(fit)
}
\arguments{
\item{fit}{An object of class \code{elicitation}.}
}
\value{
An extract from the \code{fit} object of the components used in the linear pool.
See \code{\link{fitdist}} for details.
}
\description{
Opens up a web browser (using the shiny package), from which you can choose the weights
and plot a weighted linear pool. Different distributions can be selected for the individual
experts' distribution, and feedback can be reported in the form of two quantiles from the 
weighted linear pool.
}
\details{
The weights should be positive, but do not need to some to 1; they will be normalised. 
Click the Finish button to quit the elicitation session. This will output the distributions 
and weights used in the displayed linear pool.
}
\examples{

\dontrun{
# Two experts
# Expert 1 states P(X<30)=0.25, P(X<40)=0.5, P(X<50)=0.75
# Expert 2 states P(X<20)=0.25, P(X<25)=0.5, P(X<35)=0.75
# Both experts state 0<X<100. 

v <- matrix(c(30, 40, 50, 20, 25, 35), 3, 2)
p <- c(0.25, 0.5, 0.75)
myfit <- fitdist(vals = v, probs = p, lower = 0, upper = 100)

plotWeightLP(myfit)

}
}
\author{
Jeremy Oakley <j.oakley@sheffield.ac.uk>
}
