% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivalModelExtrapolations.R
\name{survivalModelExtrapolations}
\alias{survivalModelExtrapolations}
\title{Compare Multiple Fitted Models for Survival Extrapolation}
\usage{
survivalModelExtrapolations(
  survDf,
  tEnd,
  group,
  tTruncate = NULL,
  showPlot = TRUE
)
}
\arguments{
\item{survDf}{data frame with individual patient data. Require to be a .csv file with
three columns: "time", "event" and "treatment" (in that order).
Values in the "event" column should be 0 for a censored observation, and 1 otherwise.
The"treatment" column should be included even if there is only one treatment group.'}

\item{tEnd}{the maximum time point for extrapolation}

\item{group}{character variable to select treatment group: one of the levels in the
factor variable survDf$treatment}

\item{tTruncate}{optional argument: time point at which to censor all observations}

\item{showPlot}{whether to display the plot}
}
\value{
A list containing the elements
\item{KMplot}{a ggplot2 plot object;}
\item{tMaxRange}{the time point at which there is the greatest difference between the largest
and smallest extrapolated survival probability.}
}
\description{
Fits seven parametric models to an individual patient survival data set, 
displays extrapolations, and report the time point at which there is the
widest range in estimated extrapolated survival probabilities. This is intended
to support elicitation for survival extrapolation, in particular, to inform 
the choice extrapolation time.
}
\examples{
\dontrun{

# Make a data frame using the survival::veteran data frame
sdf <- survival::veteran[, c("time", "status", "trt")]
colnames(sdf) <- c("time", "event", "treatment")
sdf$treatment <- factor(sdf$treatment, labels = c("standard", "test"))

survivalModelExtrapolations(sdf, tEnd = 100, group = "test", tTruncate = 100)
}
}
