% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elicit.R
\name{elicit}
\alias{elicit}
\title{Elicit judgements and fit distributions interactively}
\usage{
elicit()
}
\description{
Opens up a web browser (using the shiny package), from which you can specify
judgements, fit distributions and plot the fitted density functions with
additional feedback.
}
\details{
Parameter limits determine which distributions can be fitted. Non-negative
lower limits are needed for the gamma, lognormal and log-t distributions,
and both limits must be finite for to fit a beta distribution. If a
histogram is fitted without specifying finite limits, endpoints are chosen
based on fitting a normal distribution.

As an example, if the elicited judgements are P(X<15)=0.25, P(X<20)=0.5) and
P(X<40)=0.75, specify the parameter values as 15,20,40 and the cumulative
probabilities as 0.25,0.5,0.75.

Press Esc in the R console window to exit the elicitation session.
}
\examples{

\dontrun{

elicit()

}
}
\author{
Jeremy Oakley <j.oakley@sheffield.ac.uk>
}

